pyproject.toml
[tool.poetry]
name = "slack-archiver"
version = "0.1.0"
description = ""
readme = "README.md"
authors = ["joe-yama <s1r0mqme@gmail.com>"]
license = "MIT"
packages = [
  { include = "slackarchiver" }
]
repository = "https://github.com/joe-yama/slack-archiver"

[tool.poetry.dependencies]
python = ">=3.7,<4"
python-dotenv = "^0.19.2"
slack-sdk = "^3.15.2"
fire = "^0.4.0"

[tool.poetry.dev-dependencies]
pysen = {version = "^0.10.1", extras = ["lint"]}
pytest = "^7.1.0"

[tool.poetry.scripts]
list-channels = "slackarchiver.cli:cli_list_channels"
archive-channels = "slackarchiver.cli:cli_archive_channels"

[tool.pysen]
version = "0.10"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy =true
mypy_preset = "strict"
line_length = 88
py_version = "py38"

[[tool.pysen.lint.mypy_targets]]
  paths = ["."]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
